feat(): first coordiator version, not perfect
